refactor: move python code function to its own module refactor: move error messages from local to enum with thiserror refactor: use io::Error over string
