+++ src/core/menu.rs
@@ -8,7 +8,7 @@ use std::{
     process::{Command, Stdio},
 };
 
-use super::app::{AppState, PythonError, SimulationState};
+use super::{app::{AppState, SimulationState}, python::{execute_python_code, PythonError}};
 
 impl<'n> App<'n> {
     fn draw_menu_load_csv(&mut self, ui: &Ui) {
@@ -66,22 +66,9 @@ impl<'n> App<'n> {
                             .arg("--output")
                             .arg(file_path);
 
-                        match App::<'n>::execute_python_code(&mut command) {
+                        match execute_python_code(&mut command) {
                             Ok(_) => {}
-                            Err(error) => match error {
-                                PythonError::FailedToStartProcess(err) => {
-                                    eprintln!("Failed to start Python process: {}", err)
-                                }
-                                PythonError::FailedToReadOutput(err) => {
-                                    eprintln!("Failed to read Python process output: {}", err)
-                                }
-                                PythonError::OutputNotAvailable => {
-                                    eprintln!("Python process output is not available.")
-                                }
-                                PythonError::ProcessFailed(code) => {
-                                    eprintln!("Python process failed with exit code {}", code)
-                                }
-                            },
+                            Err(err) => eprintln!("{err}"),
                         }
                     }
                 }
@@ -93,24 +80,11 @@ impl<'n> App<'n> {
                 let mut command = Command::new("python3");
                 command.arg("-c").arg(&py_code).arg("--csv");
 
-                match App::<'n>::execute_python_code(&mut command) {
+                match execute_python_code(&mut command) {
                     Ok(output) => {
                         self.simulation_state = Some(SimulationState::from_csv(output));
                     }
-                    Err(error) => match error {
-                        PythonError::FailedToStartProcess(err) => {
-                            eprintln!("Failed to start Python process: {}", err)
-                        }
-                        PythonError::FailedToReadOutput(err) => {
-                            eprintln!("Failed to read Python process output: {}", err)
-                        }
-                        PythonError::OutputNotAvailable => {
-                            eprintln!("Python process output is not available.")
-                        }
-                        PythonError::ProcessFailed(code) => {
-                            eprintln!("Python process failed with exit code {}", code)
-                        }
-                    },
+                    Err(err) => eprintln!("{err}"),
                 }
             }
